Renamed CRUKSearch to CRUKSearchkitSearchUI.
