+++ index.js
@@ -1,7 +1,7 @@
 'use strict';
 
 export { default as CRUKSearchkitSearchBox } from './src/components/search-box/CRUKSearchkitSearchBox';
-export { default as CRUKSearch } from './src/components/search/Search';
+export { default as CRUKSearchkitSearchUI } from './src/components/ui/search/CRUKSearchkitSearchUI';
 export { default as CRUKSearchNoResultsDisplay } from './src/components/noresults/NoResults';
 export { default as CRUKPagination } from './src/components/pagination/Pagination';
 export { default as CRUKSearchResult } from './src/components/result/Result';
@@ -12,4 +12,4 @@ export { default as CRUKSearchHits } from './src/components/hits/Hits';
 export { default as CRUKSearchGTM } from './src/components/gtm/GTM';
 export { default as CRUKDateRange } from './src/components/date/CRUKDateRange';
 export { default as CRUKGeoSuggest } from './src/components/geosuggest/CRUKGeoSuggest';
-export { default as CRUKSearchkitSelect } from './src/components/ui/CRUKSearchkitSelect';
+export { default as CRUKSearchkitSelect } from './src/components/ui/list/CRUKSearchkitSelect';
